requirements.txt
# Core dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# GUI
PyQt5>=5.15.0

# Document processing
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# System and performance
psutil>=5.9.0
lz4>=4.0.0

# Optional: For enhanced PDF processing
# pdfminer.six>=20221105